- Added commitRemove flag to unlinkElement

+++ mirror/rw_functions.cpp
@@ -9,14 +9,17 @@
 
 #include "rw_functions.h"
 
-bool unlinkElement(directoryElement* theElement, iNodeMap& nodeSet) {
+bool unlinkElement(directoryElement* theElement, iNodeMap& nodeSet, bool commitRemove) {
 	if (theElement->get_parent() == NULL) return false;
 	
 	if (theElement->isDirectory())
 		for (list<directoryElement>::iterator it = theElement->get_contents()->begin();it!=theElement->get_contents()->end();it++)
-			unlinkElement(&(*it),nodeSet);
+			unlinkElement(&(*it),nodeSet, commitRemove);
 	
-	remove((theElement->getPathToElement()).c_str());
+	if ((commitRemove) && (remove((theElement->getPathToElement()).c_str()))) {
+		cerr << "Could not remove " << theElement->get_name() << endl;
+		exit(1);
+	}
 	
 	if (theElement->get_node()->remove_element(theElement) == 0)
 		nodeSet.deleteNode(theElement->get_node());
